Github action and end-to-end tests repositories.
In order to run the Kubewarden end-to-end tests is necessary to have two
additional repositories. The first one stores the Github action used to
setup a Kubernetes cluster where the Kubewarden stack is installed and
the tests are run. The seconds one has the files used during the tests,
like the yaml files applied and any other additional scripts.